renamed files to MeetingData and Meetings
