+++ components/MeetingsData.jsx
@@ -1,0 +1,18 @@
+export const Meetings = [
+  {
+    title: "mock interview",
+    date: "october 26, 2022(wednesday)",
+    time: "11 am",
+    location: "wch 127",
+    backgroundColor: "bg-acm-orange",
+  },
+  {
+    title: "resume review",
+    date: "october 26, 2022(wednesday)",
+    time: "11 am",
+    location: "wch 127",
+    backgroundColor: "bg-acm-marine",
+  },
+];
+
+export default Meetings;
